ide: Add support for CHS mode, add a new drive type
The main part of this code is adding CHS support to the emulation. The support
is fairly basic but should cover a lot of use cases. Primarily it is missing
support for multiple logical mappings, although strictly speaking this is not
a required drive feature.

The new drive type is a 128MB LBA capable drive.