rc2014: add a slightly modified lib765 for our use
We can then do floppy disk emulation. It's not quite a correct emulation because
we actually have WD37C65 on the real hardware but it is close enough for FDU
to work. The current ZSDOS/CPM snapshot bombs and I need to debug that a bit
both here and on my actual hardware (where it also bombs except in FDU but
differently).