rc2014: first cut at TinyZ80 emulation
See

https://github.com/skiselev/tiny_z80

It's basically an EasyZ80 with Z84C15. We have the Z84C15 bits already and
we have the EasyZ80 bits already, the rest should be a simple matter of
plumbing.

Due to the address clash the CF adapter is modelled at 0x90. The PIO is
also modelled with an SD adapter attached.