rc2014-68hc11: first cut at flat memory emulation
In flat memory mode the 68HC11 CPU card port A is connected to A19-A16 on the
extended bus. At boot time the HC11 comes up with PA6-PA4 as 0 and PA3 as an
input. A pull down on the board keeps PA3 low if it is an input.

This model relies upon the internal (unbanked) memory to work as all the main
memory switches at once.