nc100/200: Amstrad NC series emulation
A reasonably full emulation of both systems including basic PCMCIA memory
card support and nc200 floppy support. The onboard serial is not yet
emulated and the RTC emulation is a bit minimal.

Sufficient to run the system ROMs from the real NC100 and NC200, as well as
David Given's floppy loader and to boot up Fuzix.