rc2014-z180: SPI updates
- Fix a bug where Z180 without RTC didn't update the other lines (CS etc)
- Add support for the second SPI device via the buspirate

This allows you to wire an actual real SPI device to the emulator as if it was
on the Z180 CSIO, and then use the emulator to work with the actual SPI
hardware.

The ALT line on the bus pirate is used as an emulation of the RC2014 reset line
because there is no specific extra GPIO and some of the SPI interfaces attach
themselves to the RC2014 reset line as well as they need a reset signal.