mini68k: continue with the ns202 work
With these changes it's now possible to run Fuzix (note that the binary tree
will fail because the emulator emulates a 3 wire serial link without the
flow control lines so it gets stuck because it waits wrongly for RTS/CTS
even when CRTSCTS is not enabled.

A tweaked build worked fine.