z280: import from z280emu and add our z280rc
Easier to work with for rcbus stuff within the emulator kit framework
than standalone.

For the original see: z29-e,ii by Michal Tomek
https://github.com/mtdev79/z280emu

This is a fairly straight import of the emulator core. Mostly it's
just fixing warnings so it compiles cleanly.

There are some things that would fit better if changed but there's
a lot to be said for keeping the core shared as much as possible.