ns807x: get things up and running
This gets the 807x emulator able to run the NSC Nybble BASIC. The test
environment is a simple machine with 2K RAM plus space for the ROM and
a tiny bit of magic for a top of memory ROM driving external I/O as was
supported. Ctrl-C also sets SA so causes a break to interrupt the program.