wd17xx: emulation fixes
Some littleboard ROMs expect to issue a command, nap a moment then for BUSY
to come on and they use this is part of their command handling logic. Model
this properly. This also removes (or rather replaces) the busyhack stuff that
was needed and is now properly explained.

Also correct the header reading. The 1772 seems to output one less byte of
response. With this tweaked the littleboard works again.