rcbus-6809.c: mask off extra bits in PC.
A comment in the e6809.c says that register values
can have garbage data in the bits above their register size.
For example, adding a negative relative offset for the LBRA
instruction can set bit 17 of the PC. This has no functional effect
except that the disassemly when TRACE_CPU is enabled attempts
to disassemble the wrong memory.

The fix here is to mask off the PC with 0xFFFF in the code
that prints the disassembly.