Fix Clang 19+ errors
This fixes several errors that Clang produces with -Werror -Wall
-pedantic that GCC does not trigger. This should now be buildable on
systems that use Clang as the default cc.

Implementation note: Most of these are fairly straight forward, but the
if conditional setting the value of pai->ctrl in the condition should be
validated as correct. There are also a few places where a default value
is given instead of depending on undefined behavior, however, I chose
the zero value for these, but that may not be desirable.