Fix off by one on sprite vertical position.
The TMS9918a considers the top row of sprites to begin being visible at
Y=-1. (0xFF).  This change compensates for that.

When I updated my game to position sprites accordingly, they were all
rendered correctly on real hardware and rendered incorrectly on the
emulator.  Specifically, on the emulator, sprites appeared 1 pixel too
high.

This PR simply increments the Y value of the sprite in the sprite slice
routine.